requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
pyjwt==2.10.1
cryptography==44.0.0
pytest==8.3.4
pytest-cov==6.0.0